Add the operation type to the mutator types; add a dictionary append.
Knowing which type of mutators are in use helps to favour certain
kinds of operations. In particular, if you start with a small dictionary
of words and wish to generate longer strings, using just an option
that appends to the dictionary is useful.

The dictionary append operation has been added which allows the
values from the dictionary to be strung together in increasingly
longer sequences. It doesn't offer the option of appending multiple
values from the dictionary so the operation my end up in a local
minima, but such mutators could be added in the future.