Update the requirements in setup.py and requirements.txt.
The setup.py and requirements.txt had been stripped to make them
work on both python2 and 3, without regard for keeping explicit
statements of what it works with. This has been updated to specify
the enviroment checks, so that we only install the requirements
if they're needed on a given python version.

Similarly, the python_requires has been updated to say "I'll take
any 2.x version over 2.7, OR any 3.x version over 3.5.3" which
better matches with the original requirements.