Add a harness to run all the example fuzzers.
The example fuzzers provide a useful way to see how you might use
the pythonfuzz package. However, they're no use if they don't work.
Adding a tool that is able to run them all, and report on their
status will ensure that these do not get broken in the future.

The tool is able to be configured with the number of runs that it
expects to execute, and whether it will keep any crash or timeout
files that are generated. At the end of the run of each example,
it will report the collected information from the run. This should
be useful for checking the performance in the future, as well.

If any of the examples fails to function - ie it reports an error
itself, then the runner will exit with a non-0 return code. This
will allow it to be used as a gate for the examples being functional.