Update the Makefile to invoke the examples; add a levels concept to tests.
The Makefile's default target has been updated so that it now requires
that the 'system' level tests build. This is one of 3 new targets that
have been added:

  - test_level_unittests, which should cause the unit tests to run.
    We don't have any yet, but some will be added shortly.

  - test_level_integration, which causes the integration tests to run.
    These are the tests in the 'tests' directory, which exercise the
    external interface in a test-like environment to see that it
    functions as desired.
    This level is dependant on the unittests passing before it will
    be run.

  - test_level_system, which causes the examples to be run.
    These examples are the sorts of programs that a user might write,
    and which we expect to be invoked from the command and to do
    their job.
    This level is dependant on the integration tests passing before
    it will be run.

In CI, the individual test targets (rather than the levels) would
probably be invoked, with each one still dependant on the earlier ones
in stages (or maybe just run them all in parallel, as they're all very
fast).