Fix for the dictionary text insert leaving NUL characters in string.
The dictionary text insert wasn't working properly, despite looking
just fine. I think the reason is that the parameters on 'copy' are
transposed. The parameter called 'src' is actually the destination,
(and vice-versa). This caused much confusion when trying to work out
what was happening there.

This change fixes the behaviour of the dictionary insert, but defers
fixing the parameter names until it can be confirmed what was meant
and that the strings are being manipulated correctly.

The change also corrects a mistaken format character in the mutator
__repr__ method.