Add annotations to each of the test systems to describe themselves.
The test files have now been updated to include annotations, in the
file prologue comment, which describe the test and its place in the
testing environment.

These are just a convention that I've used previously, but they help
to focus anyone doing testing on describing where they fit into the
testing of the system. Such descriptions make it obvious when reviewed
where there are gaps in testing.